mini_projet/health_service/requirements.txt
flask
requests
Pyjwt